Fix ESLint and Prettier issues
