+++ packages/ui/.prettierignore
@@ -1,0 +1,2 @@
+out/
+.next/
