+++ packages/ui/jest.config.js
@@ -9,4 +9,4 @@ module.exports = {
   moduleNameMapper: {
     '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
   },
-};
+}
