+++ packages/ui/tsconfig.json
@@ -29,6 +29,7 @@
     "jest.config.js",
     "tailwind.config.js",
     "postcss.config.js",
+    "next.config.js",
     "styleMock.js"
   ],
   "exclude": [
